Initial work on MD3 importing
Add read methods for each MD3 data structure

Also, rename MD3Vertex.encode to encode_normal, and rewrite it so it's easier to read

NOTE: Right now, the importer just takes the data and prints it into a Blender text resource. There's still a lot of work to be done on the importer.