Remove tri_index and rename surftri_start
tri_index is useless, since it was only used to calculate the index of the triangle on the surface. This isn't necessary, since a running count of all triangles on all surfaces can be stored in another variable, and the running count can be used as a subscript for bl_mesh.polygons.